Update aboutus.html
Tweaking the about.us html to make it work with the new css and also wanting to make each HTML have a similar style so its easier to customize in css. I added things like the div classes and the footer. I removed the text because I will also change that.